refactorizando list afecciones separando responsabilidades en nuevos hooks y componentes
