+++ frontend/AgroSoft_f/src/modules/Sanidad/api/afeccionesApi.ts
@@ -6,6 +6,10 @@ export const getAfecciones = async ():Promise<Afecciones[]> => {
     return response.data
 };
 
+export const postAfecciones = async (data?:any):Promise<Afecciones> => {
+    const response = await apiClient.post<Afecciones>('plaga/',data);
+    return response.data
+}
 
 export const patchAfecciones = async ( id: number, data: Partial<Afecciones>): Promise<Afecciones> => {
     const response = await apiClient.patch<Afecciones>(`plaga/${id}/`, data);
