+++ frontend/AgroSoft_f/src/modules/Sanidad/hooks/afecciones/useEditarAfeccion.ts
@@ -1,0 +1,20 @@
+import { useState } from "react";
+import { UseModal } from "@/hooks/useModal";
+import { Afecciones } from "../../types";
+
+export const useEditarAfeccion = () => {
+  const { isOpen, openModal, closeModal } = UseModal();
+  const [afeccionEditada, setAfeccionEditada] = useState<Afecciones | null>(null);
+
+  const handleEditar = (afeccion: Afecciones) => {
+    setAfeccionEditada(afeccion);
+    openModal();
+  };
+
+  return {
+    isOpen,
+    closeModal,
+    afeccionEditada,
+    handleEditar,
+  };
+};
